Fix #1
Issue: Despite passing 'videoEmbeddable' as 'true' in the youtube_api_request, the search API still returns some videos that are not embeddable.

Cause: This is presumably due to video authors permitting embedding, but limiting the permission to certain sites. YouTube still considers these videos as 'embeddable', even though these videos are not embeddable for the purpose of the general public, i.e. it is not 'truly' embeddable.

Fix: The response to a HTML request for the YouTube video's embed link has the 'robots' meta tag set to 'noindex' if the video is not embeddable, while the video's that are embeddable do not have this meta tag. This difference is used to verify if the video is 'truly' embeddable.

In this commit:

- Create 'verify' view as function to verify if the video is 'truly' embeddable

- Create function to call 'verify' view from client-side JavaScript using AJAX

- Add and style 'error-message-panel' which is used to display error message if the video is not embeddable

- Also, turn off autoescape for title element of the video-item, as this leads to diplay of single quotes as HTML escape characters